Remove auth.restTokenPrivateKey from secure.config template
Support for token based REST authentication was removed in November
2012 [1] and the auth.restTokenPrivateKey setting has not been used
since version 2.6.  I'm not sure if it was even used in earlier
released versions.

The initialization of auth.restTokenPrivateKey was missed in that
removal, so a random value is still generated and set into the config
file when running the `init` program. It is not, however, used by
Gerrit anywhere.

[1] https://gerrit-review.googlesource.com/#/c/39283/

Change-Id: I84d40a8e5bb923bd50d1105217ba82afe9bd5da0