Fixes #3: Remove gerrit.sh template
The gerrit.sh file is included in the gerrit.war and gets put into
the $site/bin directory when the init command is run.

The template doesn't use any variables and apart from a couple of
whitespace diffs is the same as in the latest master branch of Gerrit.

By using a template, any changes made in new releases of Gerrit will
not be taken unless the template is also updated.

Remove the template so that the file gets installed from gerrit.war
instead.

Change-Id: Ia877a534220e53470d2d9732166b4dcc9bac15c2